+++ src/app/(user)/signup/page.tsx
@@ -1,19 +1,17 @@
 "use client";
 
-import { useLayout } from "@/components/layouts/provider/LayoutProvider";
 import { Button } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
 import { signupSchema, SignupValues } from "@/lib/schemas/user.schema";
 import { zodResolver } from "@hookform/resolvers/zod";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-import { FieldErrors, useForm } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import { UserFormField } from "../_components/UserFormField";
 import { toast } from "sonner";
+import { useUserFormError } from "@/hooks/useUserFormError";
 
 export default function JoinPage() {
-  const { setHeaderTitle } = useLayout();
   const router = useRouter();
 
   const {
@@ -34,43 +32,23 @@ export default function JoinPage() {
     },
   });
 
+  const { showErrors } = useUserFormError<SignupValues>({
+    title: "입력 정보를 확인해주세요",
+    duration: 5000,
+  });
+
   const onSubmit = (values: SignupValues) => {
     console.log(values);
     toast.success("회원가입 성공");
     router.push("/login");
   };
 
-  const onError = (errors: FieldErrors<SignupValues>) => {
-    const errorMessages = Object.values(errors)
-      .map((error) => error?.message)
-      .filter(Boolean);
-
-    if (errorMessages.length > 0) {
-      toast.error("입력 정보를 확인해주세요", {
-        description: (
-          <ul className="mt-2 space-y-1">
-            {errorMessages.map((message, index) => (
-              <li key={index} className="text-sm">
-                • {message}
-              </li>
-            ))}
-          </ul>
-        ),
-        duration: 5000,
-      });
-    }
-  };
-
-  useEffect(() => {
-    setHeaderTitle("회원가입");
-  }, [setHeaderTitle]);
-
   return (
     <div className="flex items-center justify-center p-6 pt-14">
       <div className="flex flex-col items-center w-full max-w-sm gap-4">
         <div className="text-2xl font-bold">회원가입</div>
 
-        <form className="flex flex-col w-full gap-4" onSubmit={handleSubmit(onSubmit, onError)}>
+        <form className="flex flex-col w-full gap-4" onSubmit={handleSubmit(onSubmit, showErrors)}>
           <UserFormField label="이메일" type="email" register={register("email")} placeholder="email@mail.com" error={errors.email?.message} required />
           <UserFormField label="이름" type="text" register={register("userName")} placeholder="홍길동" error={errors.userName?.message} required />
           <UserFormField label="개인번호" type="tel" register={register("userPhone")} placeholder="010-0000-0000" error={errors.userPhone?.message} required />
